lan_scanner/requirements.txt
flask
python-nmap
scapy
psutil
netifaces
mysql-connector-python
